requirements.txt
ninja  # For faster builds.
psutil
ray >= 2.5.1
pandas  # Required for Ray data.
pyarrow  # Required for Ray data.
sentencepiece  # Required for LLaMA tokenizer.
numpy
torch == 2.3
transformers >= 4.37.0  # Required for Mistral.
matplotlib
plotly_express
seaborn
wandb
kaleido==0.2.1
ddsketch
jupyterlab
flashinfer >= 0.0.5
pillow
tiktoken
grpcio
uvicorn
fastapi
openai
